# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API For News",
        "version": "1.0.0"
    },
    "paths": {
        "/api/news?page={id}": {
            "get": {
                "tags": [
                    "/news"
                ],
                "summary": "View a list of news",
                "description": "returns a list of news",
                "operationId": "a04fa10877ac5ff727b81adc4831c0d4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of news",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/news/article/{articleId}": {
            "get": {
                "tags": [
                    "/news"
                ],
                "summary": "View a list of news",
                "operationId": "getArticle",
                "parameters": [
                    {
                        "name": "articleId",
                        "in": "path",
                        "description": "ID of article",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    }
}